requirements.txt
streamlit>=1.20.0
pandas>=2.0.0
ruamel.yaml>=0.17.0
openpyxl>=3.0.0
streamlit-aggrid==1.0.3
rapidfuzz>=3.0.0